# added Mapper methods and granularity for marshalling