--- a/xstream/src/java/com/thoughtworks/xstream/core/AbstractReferenceMarshaller.java
+++ b/xstream/src/java/com/thoughtworks/xstream/core/AbstractReferenceMarshaller.java
@@ -47,7 +47,7 @@ public abstract class AbstractReferenceMarshaller extends TreeMarshaller impleme
 
     public void convert(Object item, Converter converter) {
 
-        if (getMapper().isImmutableValueType(item.getClass(), /*includeBackwardCompatibleTypes*/true)) {
+        if (getMapper().isImmutableValueType(item.getClass())) {
             // strings, ints, dates, etc... don't bother using references.
             converter.marshal(item, writer, this);
         } else {