… xstream/.idea/libraries/Maven__junit_junit_3_8_1.xml
… xstream/src/java/com/thoughtworks/xstream/core/ReferenceByXPathMarshallingStrategy.java
… xstream/src/java/com/thoughtworks/xstream/mapper/DefaultMapper.java
… xstream/src/java/com/thoughtworks/xstream/mapper/EnumMapper.java
… xstream/src/java/com/thoughtworks/xstream/mapper/ImmutableTypesMapper.java
… xstream/src/java/com/thoughtworks/xstream/mapper/Mapper.java
… xstream/src/java/com/thoughtworks/xstream/mapper/MapperWrapper.java
… xstream/src/test/com/thoughtworks/acceptance/ImmutableRegistryTest.java
… xstream/src/test/com/thoughtworks/acceptance/XStream12CompatibilityTest.java
… xstream/src/test/com/thoughtworks/xstream/core/ReferenceByXPathMarshallingStrategyTest.java