Simplified my previous implementation such that it doesn't use enums and doesn't modify any public facing API's.

To properly preserve backwards compatibility, my new method must delegate to the extra-arg free one, such that the chain of wrappers, on a call to the new method, also includes calling the old method. Since the method is looking for any positive result (ie, once anybody returns true in any context we're done), the solution to this was simply to have

```Java
public boolean isImmutableType(Class typeName, boolean includeBackwardsCompatible){
    return isImmutableType(typeName) || wrapped.isImmutableType(typeName, includeBackwardsCompatible)
}
```

This has the effect of invoking any user-written backwards-compatible-agnostic code to look for a result, and then look through the defaults in the proper scope.

The one issue I have with this implementation is that the performance of the `Mapper.isImmutableType(Class, boolean)` is porportional to the _square_ of the number of wrappers.

For example, if we have:
- StringeyMapper extends MapperWrapper 
- DomainishMapper extends MapperWrapper
- ImmutableMapper extends MapperWrapper

and we get them wrapped in the sequence

`StringeyMapper` delegates to `DomainishMapper` delegates to `ImmutableMapper` delegates to `defaultMapper`

then the call stack would be

caller | call-site
--- | ---
_consumer_ | `stringyMapper.isImmutableType(t, true)`
stringyMapper | `stringyMapper.isImmutableType(t)`
stringyMapper | `domainishMapper.isImmutableType(t)`
domainishMapper | `immutableMapper.isImmutableType(t)`
immutableMapper | `defaultMapper.isImmutableType(t)`
immutableMapper | 1-arg method on defaultMapper returns
domainishMapper | 1-arg method on immutableMapper returns
stringyMapper | 1-arg method on domainishMapper returns
stringyMapper | `domainishMapper.isImmutableType(t, true)`
domainishMapper | **redundantly** `domainishMapper.isImmutableType(t)`
domainishMapper | **redundantly** `immutableMapper.isImmutableType(t)`
immutableMapper | **redundantly** `defaultMapper.isImmutableType(t)`
immutableMapper | 1-arg method on defaultMapper returns
domainishMapper | 1-arg method on immutableMapper returns
domainishMapper | `immutableMapper.isImmutableType(t, true)`
immutableMapper | **redundantly** `immutableMapper.isImmutableType(t)`
immutableMapper | **redundantly** `defaultMapper.isImmutableType(t)`
immutableMapper | 1-arg method on defaultMapper returns
immutableMapper | defaultMapper.isImmutableType(t, true);
immutableMapper | 2-arg method on defaultMapper returns
domainishMapper | 2-arg method on immutableMapper returns
stringeyMapper | 2-arg method on domainishMapper returns
_consumer_ | 2-arg method on stringeyMapper returns 

If I'm not mistaken this should be equal to numWrappers ^2  /2, but 13 != 8 so I might be mistaken.

This means that any user-written code for Types must be re-entrant, which was not a requirement before (though no sane implementation wouldn't be). Further it means a user that puts a break-point inside his `isIMmutableType` overload is going to see it hit _alot_ more frequently, though that doesn't seem like too much of a concern to me.

A bigger concern to me was performance: all those extra calls might make a negative impact on performance. To test this I created an array list of 1,000,000 ints, and implemented a solution that kept a map of currently-inspected types which it referenced, so as to remove the redundant calls. This implementation was slightly slower than the implementation that had the redundant calls, so performance appears not to be an issue, unless I made an error in writing the performance test.

using the test
```Java
        ArrayList<Integer> ints = new ArrayList<>();
        for(int i = 0; i < 1_000_000; i++){
            ints.add(i);
        }

        StopWatch watch = new StopWatch();
        watch.start();
        String document = xstream.toXML(ints);
        System.out.println("took " + watch.getTime() + "ms to serialize");
        watch.stop();
        watch.reset();
        watch.start();

        Object result = xstream.fromXML(document);
        System.out.print("took " + watch.getTime() + "ms to deserialize");
```

and the following to address the problem
```Java
//as a static field --ThreadLocal's have been around since java 1.2, and anonymous classes since 1.1
private static final ThreadLocal<Set<Class>> inspectingTypes = new ThreadLocal<Set<Class>>(){
    protected Set<Class> initialValue() { return new HashSet<>(); }
};

//in the isImmutableType(Class, boolean) method
Set<Class> classCallStack = inspectingTypes.get();

if( ! classCallStack.contains(type)) try {
    classCallStack.add(type);

    return isImmutableValueType(type)
            || wrapped.isImmutableValueType(type, includeBackwardsCompatibleTypes);
}
finally{
    classCallStack.remove(type);
}
else{
    return wrapped.isImmutableValueType(type, includeBackwardsCompatibleTypes);
}
```

I got

scenario | serialization time | deserialization time
--- | --- | ---
with the problem | 931ms | 1720ms
without the problem | 928ms | 1672ms
with the wierd stack-set fix | 948ms | 1722ms

which to me means its all a wash, thus probably just accepting these extra calls is OK.