@Groostav 2015-04-29
I've completely screwed this up. Need to work on my boolean algebra.