@Groostav 2015-04-29
OK, so I added a test to proove that the unmarshalling class does keep references when in backwards compatible mode, which was the problem I believe I was having.