@joehni 2015-04-29
Hi Geoff, IMHO it is still too complicated. Let's reiterate for a moment:

Basically we want to avoid the memory consuming references to the unmarshalled objects. Nothing changes while marshalling. The only problem is that XML written with older versions of XStream may contain such a reference and cannot be unmarshalled anymore. Therefore we need a configuration, which types have to be referenced at unmarshalling time.

The only question of the unmarshaller is now, do I have to keep a reference? The only way to access the configuration is the Mapper, so it requires a (new) method isReferenceable(Class type).
With the ImmutableMapper you've found the best place to keep this information. And we need a way to set the flag, e.g. by adding an overloaded addImmutableType(Class type, boolean referenceable) to the XStream facade.

I do not see any necessity that the new method of the mapper is connected with the method isImmutableType. And the only one requesting the information is currently the AbstractReferenceUnmarshaller. All we have to do is a default for XStream.addImmutableType with the flag set to true (i.e. type is referenceable) and add any type in setupImmutableTypes that was already declared immutable in XStream version < 1.4.0 with the flag set to false.

BTW: XStream master requires Java 1.6. So you have to keep all the generics and @Override annotations. They can be backported in a binary compatible way for the 1.4.x branch. That is just not possible for an enum.

Regarding the benchmark: Something like StopWatch is useless for code that is executed within fractions of milliseconds. Look at [JMH](http://openjdk.java.net/projects/code-tools/jmh/) if you want to have useful results.