@Groostav 2015-04-30
The issue with adding an `isReferentiable` method is that if somebody has a complex behavior for `isImmutable` that they expect to have executed on deserialization, it wont be.

In our case, even a behavior as simple as 
```
@Override
public boolean isImmutableType(Class type){
    return type.isAnnotationPresent(com.myCompany.ImmutableAnnotation.class)
        || super.isImmutable(type);
}
```

With this strategy users would then have to add

```
@Override
public boolean isReferentiable(Class type){
    return isImmutableType(type);
}
```

which is a surprise I wouldn't want to expose anyone too, and is effectively what I'm asking the library to do by default.

That said, whats more unexpected is that forking behaviour I mentioned above.