@Groostav 2015-04-30
actually I'm being really silly, from a boolean algebra view this is quite simple.

Will do.