@Groostav 2015-04-30
Alright, so I think I'm done tell me what you think @joehni.

I dont know if github has a way to generate diffs, but I can make them easily enough and send them to you on the mailing list easily enough.

I'm not sure whats involved in your back porting project, but I figured if you could tolerate some `@Override` annotations then you would be able to tolerate all the methods on the `MapperWrapper` being tagged with them, since I hate the idea of 2 of the overriding methods being tagged with it and the rest not being tagged with it.