@Groostav 2015-05-01
ahah:

regarding generating diffs and patches, checkout [thisPR.patch](./2.patch) and [thisPR.diff](./2.diff)

nifty stuff github...