@Groostav 2015-05-01
@joehni thanks for your patience in dealing with my bumbling around an obvious solution. Are you going to apply this to the latest build (and can I close this issue?)