@joehni 2015-05-04
Hi Geoff,

when I look at your diff, then I wonder, what did you use as parent for your XStream fork? It does not seem to be the current code base... ?!?