@Groostav 2015-05-05
I  downloaded source + binaries from [XStream/downloads "source distribution"](http://xstream.codehaus.org/download.html) of the stable version