@Groostav 2015-05-05
I'll rebase that stuff into [a proper fork](../../xstream)